Moving DefaultRegionNode back as it is AWS Explorer-specific
