+++ src/lambda/lambdaTreeDataProvider.ts
@@ -14,8 +14,6 @@ import { registerCommand } from '../shared/telemetry/telemetryUtils'
 import { RefreshableAwsTreeProvider } from '../shared/treeview/awsTreeProvider'
 import { AWSCommandTreeNode } from '../shared/treeview/nodes/awsCommandTreeNode'
 import { AWSTreeNodeBase } from '../shared/treeview/nodes/awsTreeNodeBase'
-// tslint:disable-next-line:ordered-imports
-import { DefaultRegionNode } from '../shared/treeview/nodes/defaultRegionNode'
 import { ErrorNode } from '../shared/treeview/nodes/errorNode'
 import { RegionNode } from '../shared/treeview/nodes/regionNode'
 import { showErrorDetails } from '../shared/treeview/webviews/showErrorDetails'
@@ -25,6 +23,8 @@ import { deleteCloudFormation } from './commands/deleteCloudFormation'
 import { deleteLambda } from './commands/deleteLambda'
 import { invokeLambda } from './commands/invokeLambda'
 import { CloudFormationStackNode } from './explorer/cloudFormationNodes'
+// tslint:disable-next-line:ordered-imports
+import { DefaultRegionNode } from './explorer/defaultRegionNode'
 import { FunctionNodeBase } from './explorer/functionNode'
 import { StandaloneFunctionNode } from './explorer/standaloneNodes'
 import { configureLocalLambda } from './local/configureLocalLambda'
